refactor(api): inline data functions into workflow steps
Remove the data layer indirection by inlining Prisma queries directly
into workflow steps. The 12 data functions were only used within
workflows and added unnecessary abstraction.
