--- apps/api/src/data/chapters/create-chapters.ts
@@ -1,46 +1,0 @@
-import "server-only";
-import { type Chapter, prisma } from "@zoonk/db";
-import { type SafeReturn, safeAsync } from "@zoonk/utils/error";
-import { normalizeString, toSlug } from "@zoonk/utils/string";
-
-export async function createChapters(params: {
-  chapters: {
-    title: string;
-    description: string;
-  }[];
-  courseId: number;
-  language: string;
-  organizationId: number;
-}): Promise<SafeReturn<Pick<Chapter, "id" | "slug" | "title" | "description" | "position">[]>> {
-  const chaptersData = params.chapters.map((chapter, index) => ({
-    courseId: params.courseId,
-    description: chapter.description,
-    generationStatus: "pending" as const,
-    isPublished: true,
-    language: params.language,
-    normalizedTitle: normalizeString(chapter.title),
-    organizationId: params.organizationId,
-    position: index,
-    slug: toSlug(chapter.title),
-    title: chapter.title,
-  }));
-
-  const { data, error } = await safeAsync(() =>
-    prisma.chapter.createManyAndReturn({
-      data: chaptersData,
-      select: {
-        description: true,
-        id: true,
-        position: true,
-        slug: true,
-        title: true,
-      },
-    }),
-  );
-
-  if (error) {
-    return { data: null, error };
-  }
-
-  return { data, error: null };
-}
