--- apps/api/src/data/chapters/get-chapter-for-generation.ts
@@ -1,29 +1,0 @@
-import "server-only";
-import { prisma } from "@zoonk/db";
-
-export async function getChapterForGeneration(chapterId: number) {
-  return prisma.chapter.findUnique({
-    select: {
-      _count: {
-        select: {
-          lessons: true,
-        },
-      },
-      course: {
-        select: {
-          slug: true,
-          title: true,
-        },
-      },
-      description: true,
-      generationRunId: true,
-      generationStatus: true,
-      id: true,
-      language: true,
-      organizationId: true,
-      slug: true,
-      title: true,
-    },
-    where: { id: chapterId },
-  });
-}
