+++ apps/api/src/workflows/chapter-generation/steps/add-lessons-step.ts
@@ -1,4 +1,6 @@
-import { createLessons } from "@/data/lessons/create-lessons";
+import { prisma } from "@zoonk/db";
+import { safeAsync } from "@zoonk/utils/error";
+import { normalizeString, toSlug } from "@zoonk/utils/string";
 import { streamStatus } from "../stream-status";
 import { type GeneratedLesson } from "./generate-lessons-step";
 import { type ChapterContext } from "./get-chapter-step";
@@ -11,12 +13,20 @@ export async function addLessonsStep(input: {
 
   await streamStatus({ status: "started", step: "addLessons" });
 
-  const { error } = await createLessons({
+  const lessonsData = input.lessons.map((lesson, index) => ({
     chapterId: input.context.id,
+    description: lesson.description,
+    generationStatus: "pending" as const,
+    isPublished: true,
     language: input.context.language,
-    lessons: input.lessons,
+    normalizedTitle: normalizeString(lesson.title),
     organizationId: input.context.organizationId,
-  });
+    position: index,
+    slug: toSlug(lesson.title),
+    title: lesson.title,
+  }));
+
+  const { error } = await safeAsync(() => prisma.lesson.createMany({ data: lessonsData }));
 
   if (error) {
     await streamStatus({ status: "error", step: "addLessons" });
