+++ apps/api/src/workflows/chapter-generation/steps/set-chapter-as-completed-step.ts
@@ -1,6 +1,7 @@
-import { updateChapterGenerationStatus } from "@/data/chapters/update-chapter-generation-status";
 import { revalidateMainApp } from "@zoonk/core/cache/revalidate";
+import { prisma } from "@zoonk/db";
 import { cacheTagChapter } from "@zoonk/utils/cache";
+import { safeAsync } from "@zoonk/utils/error";
 import { streamStatus } from "../stream-status";
 import { type ChapterContext } from "./get-chapter-step";
 
@@ -12,11 +13,16 @@ export async function setChapterAsCompletedStep(input: {
 
   await streamStatus({ status: "started", step: "setChapterAsCompleted" });
 
-  const { error } = await updateChapterGenerationStatus({
-    chapterId: input.context.id,
-    generationRunId: input.workflowRunId,
-    generationStatus: "completed",
-  });
+  const { error } = await safeAsync(() =>
+    prisma.chapter.update({
+      data: {
+        generationRunId: input.workflowRunId,
+        generationStatus: "completed",
+      },
+      select: { generationStatus: true, id: true },
+      where: { id: input.context.id },
+    }),
+  );
 
   if (error) {
     await streamStatus({ status: "error", step: "setChapterAsCompleted" });
