+++ apps/api/src/workflows/course-generation/steps/add-categories-step.ts
@@ -1,4 +1,5 @@
-import { addCourseCategories } from "@/data/courses/add-course-categories";
+import { prisma } from "@zoonk/db";
+import { safeAsync } from "@zoonk/utils/error";
 import { streamStatus } from "../stream-status";
 import { type CourseContext } from "../types";
 
@@ -10,10 +11,17 @@ export async function addCategoriesStep(input: {
 
   await streamStatus({ status: "started", step: "addCategories" });
 
-  const { error } = await addCourseCategories({
-    categories: input.categories,
+  const categories = input.categories.map((category) => ({
+    category,
     courseId: input.course.courseId,
-  });
+  }));
+
+  const { error } = await safeAsync(() =>
+    prisma.courseCategory.createMany({
+      data: categories,
+      skipDuplicates: true,
+    }),
+  );
 
   if (error) {
     await streamStatus({ status: "error", step: "addCategories" });
