+++ apps/api/src/workflows/course-generation/steps/add-chapters-step.ts
@@ -1,4 +1,6 @@
-import { createChapters } from "@/data/chapters/create-chapters";
+import { prisma } from "@zoonk/db";
+import { safeAsync } from "@zoonk/utils/error";
+import { normalizeString, toSlug } from "@zoonk/utils/string";
 import { streamStatus } from "../stream-status";
 import { type CourseContext, type CreatedChapter, type GeneratedChapter } from "../types";
 
@@ -10,12 +12,31 @@ export async function addChaptersStep(input: {
 
   await streamStatus({ status: "started", step: "addChapters" });
 
-  const { data: createdChapters, error } = await createChapters({
-    chapters: input.chapters,
+  const chaptersData = input.chapters.map((chapter, index) => ({
     courseId: input.course.courseId,
+    description: chapter.description,
+    generationStatus: "pending" as const,
+    isPublished: true,
     language: input.course.language,
+    normalizedTitle: normalizeString(chapter.title),
     organizationId: input.course.organizationId,
-  });
+    position: index,
+    slug: toSlug(chapter.title),
+    title: chapter.title,
+  }));
+
+  const { data: createdChapters, error } = await safeAsync(() =>
+    prisma.chapter.createManyAndReturn({
+      data: chaptersData,
+      select: {
+        description: true,
+        id: true,
+        position: true,
+        slug: true,
+        title: true,
+      },
+    }),
+  );
 
   if (error || !createdChapters) {
     await streamStatus({ status: "error", step: "addChapters" });
