+++ apps/api/src/workflows/course-generation/steps/get-ai-organization-step.ts
@@ -1,9 +1,13 @@
-import { getAIOrganization } from "@/data/orgs/get-ai-organization";
+import { prisma } from "@zoonk/db";
+import { AI_ORG_SLUG } from "@zoonk/utils/constants";
 
 export async function getAIOrganizationStep(): Promise<{
   id: number;
 }> {
   "use step";
 
-  return getAIOrganization();
+  return prisma.organization.findUniqueOrThrow({
+    select: { id: true },
+    where: { slug: AI_ORG_SLUG },
+  });
 }
