+++ apps/api/src/workflows/course-generation/steps/get-course-chapters-step.ts
@@ -1,4 +1,5 @@
-import { getCourseChapters } from "@/data/chapters/get-course-chapters";
+import { prisma } from "@zoonk/db";
+import { safeAsync } from "@zoonk/utils/error";
 import { streamStatus } from "../stream-status";
 import { type CreatedChapter } from "../types";
 
@@ -7,7 +8,19 @@ export async function getCourseChaptersStep(courseId: number): Promise<CreatedCh
 
   await streamStatus({ status: "started", step: "getExistingChapters" });
 
-  const { data: chapters, error } = await getCourseChapters(courseId);
+  const { data: chapters, error } = await safeAsync(() =>
+    prisma.chapter.findMany({
+      orderBy: { position: "asc" },
+      select: {
+        description: true,
+        id: true,
+        position: true,
+        slug: true,
+        title: true,
+      },
+      where: { courseId },
+    }),
+  );
 
   if (error || !chapters) {
     await streamStatus({ status: "error", step: "getExistingChapters" });
