+++ apps/api/src/workflows/course-generation/steps/update-course-step.ts
@@ -1,4 +1,5 @@
-import { updateAICourse } from "@/data/courses/update-ai-course";
+import { prisma } from "@zoonk/db";
+import { safeAsync } from "@zoonk/utils/error";
 import { streamStatus } from "../stream-status";
 import { type CourseContext } from "../types";
 
@@ -11,12 +12,16 @@ export async function updateCourseStep(input: {
 
   await streamStatus({ status: "started", step: "updateCourse" });
 
-  const { error } = await updateAICourse({
-    courseId: input.course.courseId,
-    description: input.description,
-    generationStatus: "completed",
-    imageUrl: input.imageUrl ?? undefined,
-  });
+  const { error } = await safeAsync(() =>
+    prisma.course.update({
+      data: {
+        description: input.description,
+        generationStatus: "completed",
+        ...(input.imageUrl && { imageUrl: input.imageUrl }),
+      },
+      where: { id: input.course.courseId },
+    }),
+  );
 
   if (error) {
     await streamStatus({ status: "error", step: "updateCourse" });
